tag/decoder_test.go
//
// Copyright (c) 2018- yutopp (yutopp@gmail.com)
//
// Distributed under the Boost Software License, Version 1.0. (See accompanying
// file LICENSE_1_0.txt or copy at  https://www.boost.org/LICENSE_1_0.txt)
//

package tag

import (
	"bytes"
	"github.com/stretchr/testify/assert"
	"testing"
)

func TestDecodeAudioDataCommon(t *testing.T) {
	for _, tc := range audioDataTestCases {
		tc := tc // capture

		t.Run(tc.Name, func(t *testing.T) {
			t.Parallel()

			buf := bytes.NewBuffer(tc.Binary)
			audioData, err := DecodeAudioData(buf)
			assert.Nil(t, err)
			assert.Equal(t, tc.Value, audioData)
		})
	}
}

func TestDecodeVideoDataCommon(t *testing.T) {
	for _, tc := range videoDataTestCases {
		tc := tc // capture

		t.Run(tc.Name, func(t *testing.T) {
			t.Parallel()

			buf := bytes.NewBuffer(tc.Binary)
			videoData, err := DecodeVideoData(buf)
			assert.Nil(t, err)
			assert.Equal(t, tc.Value, videoData)
		})
	}
}
